filehandler/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pillow
python-docx